modified post_gen_script

+++ hooks/post_gen_project.py
@@ -12,10 +12,12 @@ def run_post_gen_hook(install_path="{{cookiecutter.install_path}}", folder_name=
     
     # Copy files to the installation path
     # Modify this part according to your setup requirements
-    shutil.copytree('./', os.path.join(install_path, folder_name))
-    # shutil.rmtree('./')
-    cwd = os.getcwd()
-    shutil.rmtree(cwd)
+    if(install_path=='./'):
+        pass
+    else:
+        shutil.copytree('./', os.path.join(install_path, folder_name))
+        cwd = os.getcwd()
+        shutil.rmtree(cwd)
     # print(os.listdir())
     print(f"Setup completed in {install_path}{folder_name}")
